% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_package.R
\docType{package}
\name{RcmdrPlugin.BioStat.old}
\alias{RcmdrPlugin.BioStat.old}
\alias{RcmdrPlugin.BioStat.old-package}
\title{An R Commander Plugin for Basic (Bio)Statistical Routines (old version)}
\description{
\code{RcmdrPlugin.BioStat.old} is an \code{Rcmdr} plugin for functions in package
\code{BioStat.old} and other common (bio)statistical routines.
}
